requirements.txt
astroid==3.0.0
black==23.11.0
blurhash==1.1.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==5.1.1
dill==0.3.7
idna==3.4
isort==5.12.0
Mastodon.py==1.8.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
pylint==3.0.1
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.31.0
six==1.16.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==2.0.7
